+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added option to hide "There are blocks in the way!" message
 - Added option to hide Combo messages (Won't hide the MF ones)
 - Added Zoom on etherwarp (I just really missed that)
+- Added option to make the health bar blue when the mob is shurikened (On by default)
 
 ### Fixes:
 - Made flare alert not proc when a firework is active
@@ -11,3 +12,4 @@
 ### Back-end:
 
 ### Other:
+- Added a star to the health bar whenever the mob is shurikened